Update Model User, Company

+++ src/main/java/vn/ifine/jobhunter/controller/admin/UserController.java
@@ -2,6 +2,8 @@ package vn.ifine.jobhunter.controller.admin;
 
 import java.util.List;
 
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -14,10 +16,17 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.turkraft.springfilter.boot.Filter;
+
 import jakarta.validation.Valid;
 import vn.ifine.jobhunter.domain.User;
 import vn.ifine.jobhunter.domain.response.ApiResponse;
+import vn.ifine.jobhunter.domain.response.ResultPaginationDTO;
+import vn.ifine.jobhunter.domain.response.user.ResCreateUserDTO;
+import vn.ifine.jobhunter.domain.response.user.ResUpdateUserDTO;
+import vn.ifine.jobhunter.domain.response.user.ResUserDTO;
 import vn.ifine.jobhunter.service.UserService;
+import vn.ifine.jobhunter.util.error.IdInvalidException;
 
 @RestController
 @RequestMapping("/api/admin")
@@ -31,34 +40,57 @@ public class UserController {
     }
 
     @PostMapping("/users")
-    public ResponseEntity<ApiResponse<User>> createUser(@Valid @RequestBody User reqUser) {
+    public ResponseEntity<ApiResponse<ResCreateUserDTO>> createUser(@Valid @RequestBody User reqUser)
+            throws IdInvalidException {
+        boolean isEmailExist = this.userService.isEmailExist(reqUser.getEmail());
+        if (isEmailExist) {
+            throw new IdInvalidException("Email " + reqUser.getEmail() + " đã tồn tại, vui lòng sử dụng email khác");
+        }
         String hashPassword = this.passwordEncoder.encode(reqUser.getPassword());
         reqUser.setPassword(hashPassword);
         User user = this.userService.handleCreateUser(reqUser);
         return ResponseEntity.status(HttpStatus.CREATED)
-                .body(ApiResponse.created("Create a new user successfully", user));
+                .body(ApiResponse.created("Create a new user successfully",
+                        this.userService.convertToResCreateUserDTO(user)));
     }
 
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<ApiResponse<Void>> deleteUser(@PathVariable("id") long id) {
+    public ResponseEntity<ApiResponse<Void>> deleteUser(@PathVariable("id") long id) throws IdInvalidException {
+        User currentUser = this.userService.fetchUserById(id);
+        if (currentUser == null) {
+            throw new IdInvalidException("User với id = " + id + " không tồn tại");
+        }
         this.userService.handleDeleteUser(id);
-        return ResponseEntity.ok(ApiResponse.<Void>success("Delete user successfully", null));
+        return ResponseEntity
+                .ok(ApiResponse.success("Delete user successfully", null));
     }
 
     @GetMapping("/users/{id}")
-    public ResponseEntity<ApiResponse<User>> getUserById(@PathVariable("id") long id) {
-        return ResponseEntity.ok(ApiResponse.success("Fetch a user successfully", this.userService.fetchUserById(id)));
+    public ResponseEntity<ApiResponse<ResUserDTO>> getUserById(@PathVariable("id") long id) throws IdInvalidException {
+        User currentUser = this.userService.fetchUserById(id);
+        if (currentUser == null) {
+            throw new IdInvalidException("User với id = " + id + " không tồn tại");
+        }
+        return ResponseEntity.ok(
+                ApiResponse.success("Fetch a user successfully", this.userService.convertToResUserDTO(currentUser)));
     }
 
     @GetMapping("/users")
-    public ResponseEntity<ApiResponse<List<User>>> getAllUser() {
-        return ResponseEntity.ok(ApiResponse.success("Fetch all user successfully", this.userService.fetchAllUser()));
+    public ResponseEntity<ApiResponse<ResultPaginationDTO>> getAllUser(@Filter Specification<User> spec,
+            Pageable pageable) {
+        return ResponseEntity
+                .ok(ApiResponse.success("Fetch all user successfully", this.userService.fetchAllUser(spec, pageable)));
     }
 
     @PutMapping("/users")
-    public ResponseEntity<ApiResponse<User>> updateUser(@RequestBody User reqUser) {
+    public ResponseEntity<ApiResponse<ResUpdateUserDTO>> updateUser(@RequestBody User reqUser)
+            throws IdInvalidException {
         User user = this.userService.handleUpdateUser(reqUser);
-        return ResponseEntity.ok(ApiResponse.success("Fetch a user successfully", user));
+        if (user == null) {
+            throw new IdInvalidException("User với id = " + reqUser.getId() + " không tồn tại");
+        }
+        return ResponseEntity
+                .ok(ApiResponse.success("Fetch a user successfully", this.userService.convertToResUpdateUserDTO(user)));
     }
 
 }
